% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colmap.R
\name{colmap}
\alias{colmap}
\title{Colmap}
\usage{
colmap(map = departamentos, data = NULL, var = NULL, map_id = "id",
  data_id = map_id, legend = TRUE, autocomplete = FALSE)
}
\arguments{
\item{map}{a \code{\link{fortify}}able object sucha as a
\code{\linkS4class{SpatialPolygonsDataFrame}} resulting in a
data.frame with columns \code{x} or \code{long}, \code{y} or \code{lat} and
\code{region} or \code{id}.}

\item{data}{a data.frame that contains an id variable to match the map
regions and the variable to plot over the map. By default no data is needed
in which case only the map is represented coloring the regions with ggplot's
default pallete.}

\item{var}{character, the name of the variable in data to be represented in
the map. By default it will take the first non id variable.}

\item{map_id}{character, name of the map id variable. By default \code{"id"}
which is how the region id is set on the spatial objects through
\code{fortify}.}

\item{data_id}{character, name of the data variable to match the map
id.}

\item{legend}{logical, include legend (color guide) in the plot? No legend
is included if only the map is ploted.}

\item{autocomplete}{logical, if there are any map regions without
corresponding data values and autocomplete is true then the data will be
auto-completed with missing values to plot all the regions instead of having
void portions in the map.}
}
\value{
a ggplot object.
}
\description{
colmap - for colombian map (or in general color map) - is a wrapper of
ggplot and its geom_map so that its easy to make a quick choropleth map. You
can get further control with ggplot.
}
\examples{
colmap() # Plot default map: Colombia departments.

}

