% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colmaps.R
\docType{package}
\name{colmaps}
\alias{colmaps}
\alias{colmaps-package}
\title{colmaps: Colombian maps data}
\description{
The colmaps package provides \code{\linkS4class{SpatialPolygonsDataFrame}}
objects of colombian municipalities and departments, useful to make
choropleth maps, namely \code{\link{municipios}} and
\code{\link{departamentos}}.
}
\details{
Data gathered from the
\href{http://dane.gov.co}{National Department of Statistics (DANE)}
\url{http://geoportal.dane.gov.co:8084/Divipola}.
}

